M4 Wave 4 I1: atk/raster/cmd strict flags; fix WriteShare dropped-height datastream bug, 4 scanf LP64 half-writes, procedure-cast dispatch table normalization, ENTER/LEAVE macro ptr-format bug
FossilOrigin-Name: 6a5fbc4c9c0224f6d8a5658785d3e324b02256e5b49d38ff883495fafef6f998

+++ src/atk/raster/cmd/dispbox.h
@@ -34,8 +34,8 @@ extern char debug;
 
 #define Debug(s) {printf s ; fflush(stdout);}
 #define DEBUG(s) {if (debug) {printf s ; fflush(stdout);}}
-#define ENTER(r) DEBUG(("Enter %s(0x%lx)\n", "r", self))
-#define LEAVE(r) DEBUG(("Leave %s(0x%lx)\n", "r", self))
+#define ENTER(r) DEBUG(("Enter %s(0x%lx)\n", "r", (unsigned long) self))
+#define LEAVE(r) DEBUG(("Leave %s(0x%lx)\n", "r", (unsigned long) self))
 
 #define DisplayAndReturn(self, String) {message_DisplayString(self, 0, String); return;}
